+++ sphinx/source/changelog.rst
@@ -18,6 +18,7 @@ Changelog
   :ref:`page` section
 * Moved the ``Logo`` parameter from the ``[Paths]`` section to the
   :ref:`ref-game` section and renamed it ``LogoImage``
+* Annotated the source code remnants at 39936 in `jet_set_willy.ctl`
 
 3.4 (2013-07-08)
 ----------------